Function for creating chart configurations.

create_config creates configuration lists using

create_query_url to build the SDMX query URLs

create_config(sdmx_data_query = stop("'sdmx_data_query' must be specified"),
  data_api_endpoint = "https://stats.oecd.org/SDMX-JSON/data",
  path = NULL, title = NULL, subtitle = NULL, source_url = NULL,
  source_label = NULL, highlight = NULL, baseline = NULL,
  chart_dimension = NULL, display = "label", width = NULL,
  height = NULL, x_axis = NULL, y_axis = NULL, logo = TRUE,
  owner = TRUE, language = "en",
  type = stop("'type' must be specified"))

  create_query_url(sdmx_data_query = stop("'sdmx_data_query' must be provided"),
  data_api_endpoint = stop("'data_api_endpoint' must be provided"))

Arguments

sdmx_data_query
complete SDMX URL, e.g. ""KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"

data_api_endpoint
SDMX API URL, e.g. http://stats.oecd.org/SDMX-JSON/data

path
path to a JSON template that can be read using jsonlite::fromJSON

title
character string main title of chart

subtitle
character vector chart subtitle

source_url
character string to add URL below chart

source_label
character string to show instead of URL below chart

highlight
a list of character vectors respecting the order of dimensions of the SDMX data flow

baseline
a list of character vectors respecting the order of dimensions of the SDMX data flow

chart_dimension
list of id, x, y and mode

width
integer chart width in pixel

height
integer chart height in pixel

logo
logical include logo

owner
logical include copyright information

language
character string, one of en or fr

type
one of BarChart, RowChart, ScatterChart, HSymbolChart, VSymbolChart, TimelineChart, StackedBarChart

Source

jsonlite was created by Jeroen Ooms et al., see https://cran.r-project.org/web/packages/jsonlite/index.html.

Examples

query <- "KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"
chartconfig <- create_config(sdmx_data_query = query, type = "BarChart")
str(chartconfig)

#> List of 3
#>  $ data :List of 3
#>   ..$ data   :List of 6
#>   .. ..$ title    : NULL
#>   .. ..$ subtitle : NULL
#>   .. ..$ footnotes:List of 2
#>   .. .. ..$ source     : NULL
#>   .. .. ..$ sourceLabel: NULL
#>   .. ..$ frequency: NULL
#>   .. ..$ series   : list()
#>   .. ..$ share    :List of 4
#>   .. .. ..$ focused       :List of 2
#>   .. .. .. ..$ highlight: list()
#>   .. .. .. ..$ baseline : list()
#>   .. .. ..$ chartDimension: NULL
#>   .. .. ..$ display       : chr "label"
#>   .. .. ..$ source        : chr "https://stats.oecd.org/SDMX-JSON/data/KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?st"| __truncated__
#>   ..$ options:List of 2
#>   .. ..$ base:List of 2
#>   .. .. ..$ width : NULL
#>   .. .. ..$ height: NULL
#>   .. ..$ axis:List of 2
#>   .. .. ..$ x:List of 1
#>   .. .. .. ..$ linear: Named list()
#>   .. .. ..$ y:List of 1
#>   .. .. .. ..$ linear: Named list()
#>   ..$ config :List of 4
#>   .. ..$ logo         : chr "https://upload.wikimedia.org/wikipedia/sco/0/0d/OECD_logo_new.svg"
#>   .. ..$ owner        : chr "©OECD"
#>   .. ..$ terms        :List of 2
#>   .. .. ..$ label: chr "Terms & Conditions"
#>   .. .. ..$ link : chr "https://www.oecd.org/termsandconditions/"
#>   .. ..$ sourceHeaders:List of 2
#>   .. .. ..$ Accept         : chr "application/vnd.sdmx.data+json"
#>   .. .. ..$ Accept-Language: chr "en"
#>  $ share: chr "latest"
#>  $ type : chr "BarChart"

## using custom template
jsonfile <- system.file("templates/default.json", package = "rcw")
chartconfig2 <- create_config(sdmx_data_query = query, type = "BarChart", path = jsonfile)
str(chartconfig2)

#> List of 3
#>  $ data :List of 3
#>   ..$ data   :List of 6
#>   .. ..$ title    : NULL
#>   .. ..$ subtitle : NULL
#>   .. ..$ footnotes:List of 2
#>   .. .. ..$ source     : NULL
#>   .. .. ..$ sourceLabel: NULL
#>   .. ..$ frequency: NULL
#>   .. ..$ series   : list()
#>   .. ..$ share    :List of 4
#>   .. .. ..$ focused       :List of 2
#>   .. .. .. ..$ highlight: list()
#>   .. .. .. ..$ baseline : list()
#>   .. .. ..$ chartDimension: NULL
#>   .. .. ..$ display       : chr "label"
#>   .. .. ..$ source        : chr "https://stats.oecd.org/SDMX-JSON/data/KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?st"| __truncated__
#>   ..$ options:List of 2
#>   .. ..$ base:List of 2
#>   .. .. ..$ width : NULL
#>   .. .. ..$ height: NULL
#>   .. ..$ axis:List of 2
#>   .. .. ..$ x:List of 1
#>   .. .. .. ..$ linear: Named list()
#>   .. .. ..$ y:List of 1
#>   .. .. .. ..$ linear: Named list()
#>   ..$ config :List of 4
#>   .. ..$ logo         : chr "https://upload.wikimedia.org/wikipedia/sco/0/0d/OECD_logo_new.svg"
#>   .. ..$ owner        : chr "©OECD"
#>   .. ..$ terms        :List of 2
#>   .. .. ..$ label: chr "Terms & Conditions"
#>   .. .. ..$ link : chr "https://www.oecd.org/termsandconditions/"
#>   .. ..$ sourceHeaders:List of 2
#>   .. .. ..$ Accept         : chr "application/vnd.sdmx.data+json"
#>   .. .. ..$ Accept-Language: chr "en"
#>  $ share: chr "latest"
#>  $ type : chr "BarChart"

query <- "KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"
create_query_url(sdmx_data_query = query,
                 data_api_endpoint = "http://stats.oecd.org:80/SDMX-JSON/data")

#> [1] "http://stats.oecd.org:80/SDMX-JSON/data/KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015&dimensionAtObservation=AllDimensions"