Function for creating rcw D3 graphs.

rcwChart creates simple D3 JavaScript force directed network graphs.

rcwChart(data, type = stop("'type' must be provided"), height = "300",
  width = "600")

rcwChartOutput(outputId, width = "100%", height = "500px")

renderrcwChart(expr, env = parent.frame(), quoted = FALSE)

Arguments

data
a data frame object with three columns. The first two are the names of the linked units. The third records an edge value. (Currently the third column doesn't affect the graph.)

type
a valid chart type e.g. "bar", "scatter", "row" etc.

height
height for the network graph's frame area in pixels (if NULL then height is automatically determined based on context)

width
numeric width for the network graph's frame area in pixels (if NULL then width is automatically determined based on context)

Source

D3.js was created by Michael Bostock. See http://d3js.org/ and, more specifically for directed networks https://github.com/mbostock/d3/wiki/Force-Layout

Examples

# Fake data
Source <- c("A", "A", "A", "A", "B", "B", "C", "C", "D")
Target <- c("B", "C", "D", "J", "E", "F", "G", "H", "I")
NetworkData <- data.frame(Source, Target)

# Create graph
rcwChart(NetworkData)

#> Error in rcwChart(NetworkData): 'type' must be provided